+++ src/main/java/com/unstoppabledomains/resolution/TickerVersion.java
@@ -1,9 +1,0 @@
-package com.unstoppabledomains.resolution;
-
-@Deprecated 
-public enum TickerVersion {
-  ERC20,
-  OMNI,
-  EOS,
-  TRON
-}