+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Remove deprecated Resolution#email, use Resolution#getEmail() instead
 * Remove deprecated Resolution#ipfsHash, use Resolution#getIpfsHash() instead
 * Remove deprecated Resolution#owner, use Resolution#getOwner() instead
+* Remove deprecated TickerVersion enum
 
 ## 2.0.0
 * remove ENS support